+++ python/cugraph/community/ecg_wrapper.pyx
@@ -16,7 +16,7 @@
 # cython: embedsignature = True
 # cython: language_level = 3
 
-cimport cugraph.community.louvain as c_louvain
+cimport cugraph.community.ecg as c_ecg
 from cugraph.structure.graph cimport *
 from cugraph.structure import graph_wrapper
 from cugraph.utilities.column_utils cimport *
@@ -30,13 +30,13 @@ import rmm
 import numpy as np
 
 
-def louvain(input_graph):
+def ecg(input_graph, min_weight=.05, ensemble_size=16):
     """
-    Call louvain
+    Call ECG
     """
     cdef uintptr_t graph = graph_wrapper.allocate_cpp_graph()
     cdef Graph * g = <Graph*> graph
-
+    
     if input_graph.adjlist:
         [offsets, indices] = graph_wrapper.datatype_cast([input_graph.adjlist.offsets, input_graph.adjlist.indices], [np.int32])
         [weights] = graph_wrapper.datatype_cast([input_graph.adjlist.weights], [np.float32, np.float64])
@@ -52,54 +52,22 @@ def louvain(input_graph):
         offsets, indices, values = graph_wrapper.get_adj_list(graph)
         input_graph.adjlist = input_graph.AdjList(offsets, indices, values)
 
-    # we should add get_number_of_vertices() to Graph (and this should be
-    # used instead of g.adjList.offsets.size - 1)
     num_verts = g.adjList.offsets.size - 1
 
     df = cudf.DataFrame()
     df['vertex'] = cudf.Series(np.zeros(num_verts, dtype=np.int32))
     cdef gdf_column c_index_col = get_gdf_column_view(df['vertex'])
     g.adjList.get_vertex_identifiers(&c_index_col)
-    
-
-    df['partition'] = cudf.Series(np.zeros(num_verts,dtype=np.int32))
-    cdef uintptr_t c_louvain_parts_ptr = get_column_data_ptr(df['partition']._column)
-
-    cdef bool single_precision = False
-    # this implementation is tied to cugraph.cu line 503
-    # cudaDataType_t val_type = graph->adjList->edge_data?
-    #     gdf_to_cudadtype(graph->adjList->edge_data): CUDA_R_32F;
-    # this is tied to the low-level implementation detail of the lower level
-    # function, and very vulnerable to low level changes. Better be
-    # reimplemented, but we are planning to eventually remove nvgraph, so I may
-    # leave as is right at this moment.
-    if g.adjList.edge_data:
-        if g.adjList.edge_data.dtype == GDF_FLOAT32:
-            single_precision = True;
-    else:
-        single_precision = True;
-
-    cdef float final_modularity_single_precision = 1.0
-    cdef double final_modularity_double_precision = 1.0
-    cdef int num_level = 0
-    
-
-    if single_precision:
-        c_louvain.louvain(<Graph*>g,
-                  <void*>&final_modularity_single_precision,
-                  <void*>&num_level, <void*>c_louvain_parts_ptr,
-                  100)
-    else:
-        c_louvain.louvain(<Graph*>g,
-                  <void*>&final_modularity_double_precision,
-                  <void*>&num_level, <void*>c_louvain_parts_ptr,
-                  100)
-    
-
     if input_graph.renumbered:
         df['vertex'] = input_graph.edgelist.renumber_map[df['vertex']]
 
-    if single_precision:
-        return df, <double>final_modularity_single_precision
+    df['partition'] = cudf.Series(np.zeros(num_verts, dtype=np.int32))
+    cdef uintptr_t c_ecg_ptr = get_column_data_ptr(df['partition']._column)
+
+    if g.adjList.edge_data.dtype == np.float32:
+        c_ecg.ecg[int32_t, float] (<Graph*>g, min_weight, ensemble_size, <int32_t*>c_ecg_ptr)
     else:
-        return df, final_modularity_double_precision
+        c_ecg.ecg[int32_t, double] (<Graph*>g, min_weight, ensemble_size, <int32_t*>c_ecg_ptr)
+
+    return df
+